requirements.txt
google-generativeai
feedparser
requests
line-bot-sdk
python-dateutil
scikit-learn
numpy
